Hide the ActionBar FAB
I originally intended for it to exist so that you could have a sort of "immersive mode" when drawing, but it actually takes up more space and can be quite distracting at times.
Plus, when you do actually end up using it, you can only use whatever tool you want to the edges of your screen. If you're using it to immersively paint (which was the intended purpose), you have to inevitably show the top bar again so you can pan -- this basically renders the button useless, especially on smaller screens.