Use Pathway to generate SVG path data
It's kinda sad that I won't be able to make my own implementation (because it's fun!) but Pathway's is much, much more fully fleshed out and just works™.
This is the best option to make FreePaint as stable and feature-rich as possible. Anyway Pathway is awesome, I recommend giving it a star.